requirements.txt
huggingface_hub
langchain_huggingface
langchain_ai21
langchain_mongodb
python-dotenv
gradio
minijinja
transformers
ctransformers
langchain
langchain_community
torch
pypdf
sentence-transformers
faiss-cpu
openai
bitsandbytes 
accelerate 
xformers 
einops 
pymongo
rank_bm25
flask
chardet